Address PR review feedback
- Extract useDebounce to reusable hook in src/hooks/useDebounce.ts
- Wrap allTags computation in useMemo to avoid recalculating on every render
- Improve generateIdentifier to handle diacritics via NFD normalization
- Add explanatory comments for setTimeout delay and MCP response structure
- Increase refresh delay to 1000ms with comment about async task processing
