+++ frontend/src/hooks/useDebounce.ts
@@ -1,0 +1,20 @@
+import { useState, useEffect } from 'react'
+
+/**
+ * Custom hook for debouncing values.
+ * Useful for delaying expensive operations like API calls until user stops typing.
+ *
+ * @param value - The value to debounce
+ * @param delay - Delay in milliseconds
+ * @returns The debounced value
+ */
+export function useDebounce<T>(value: T, delay: number): T {
+  const [debouncedValue, setDebouncedValue] = useState<T>(value)
+
+  useEffect(() => {
+    const timer = setTimeout(() => setDebouncedValue(value), delay)
+    return () => clearTimeout(timer)
+  }, [value, delay])
+
+  return debouncedValue
+}
